+++ filenames.gni
@@ -684,8 +684,8 @@ filenames = {
     "shell/common/skia_util.cc",
     "shell/common/skia_util.h",
     "shell/common/thread_restrictions.h",
-    "shell/common/v8_value_serializer.cc",
-    "shell/common/v8_value_serializer.h",
+    "shell/common/v8_util.cc",
+    "shell/common/v8_util.h",
     "shell/common/world_ids.h",
     "shell/renderer/api/context_bridge/object_cache.cc",
     "shell/renderer/api/context_bridge/object_cache.h",
