+++ shell/common/v8_util.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by the MIT license that can be
 // found in the LICENSE file.
 
-#include "shell/common/v8_value_serializer.h"
+#include "shell/common/v8_util.h"
 
 #include <utility>
 #include <vector>
